hardening(tier-2): float money columns → NUMERIC(18, 2) [14 cols, 11 models]
TIER_2 migration — explicit founder approval received 2026-04-15.

Converts 14 monetary columns from DOUBLE PRECISION to NUMERIC(18, 2)
via alembic zzz8_float_money_to_numeric with matching model flips.
Eliminates the silent float-cent rounding class at reconciliation
time — the bug shape that reconciles wrong invisibly and only
surfaces when a merchant asks "why is this invoice off by 3 cents".
At 10k merchants × millions of orders the cumulative error becomes
merchant-visible; top-1 bar will not accept.

Columns (frozen debt allowlist is now empty):
  action_snapshots.baseline_revenue_7d / delta_revenue_7d
  active_nudges.estimated_revenue_window
  ad_spend_daily.revenue_attributed_eur / spend_eur
  analytics_events.revenue_eur
  execution_baselines.product_b_revenue_24h
  price_watch.last_seen_price / previous_price
  product_metrics.revenue_24h
  scaling_recommendations.estimated_cost_increase_eur
  shop_orders.total_price   ← most critical, live order totals
  system_snapshots.llm_estimated_cost_eur
  trust_execution_log.revenue_delta_eur

Side fixes:
- scaling_intelligence.build_forecast coerces llm_estimated_cost_eur
  to float at the boundary (forecast linreg mixes ints, Python
  refuses float * Decimal).
- elite v2 money-column test pattern broadened to include _spend /
  _payout / _eur / _usd / _gbp — caught ad_spend_daily.spend_eur
  mid-migration, which was then included in the ALTER batch.
- _FLOAT_MONEY_DEBT_ALLOWLIST now empty — frozen ledger for future
  regressions.

Lock semantics: ALTER COLUMN TYPE holds ACCESS EXCLUSIVE for the
rewrite duration. shop_orders is the largest and is sequential-scan
rewritten. Production deploy must land in a maintenance window;
dev DB is already migrated (alembic current = zzz8_...).

Rollback: downgrade reverses via DOUBLE PRECISION cast. Zero data
loss because the rewritten data was originally double-precision.

+++ backend/app/models/action_snapshot.py
@@ -9,7 +9,7 @@ from __future__ import annotations
 
 from datetime import datetime
 
-from sqlalchemy import Boolean, Column, DateTime, Float, Index, Integer, String
+from sqlalchemy import Boolean, Column, DateTime, Float, Index, Integer, Numeric, String
 
 from app.core.database import Base
 
@@ -26,7 +26,7 @@ class ActionSnapshot(Base):
     # Baseline metrics at snapshot time
     baseline_cvr = Column(Float, nullable=True)
     baseline_atc_rate = Column(Float, nullable=True)
-    baseline_revenue_7d = Column(Float, nullable=True)
+    baseline_revenue_7d = Column(Numeric(18, 2), nullable=True)
     baseline_visitors_7d = Column(Integer, nullable=True)
     baseline_orders_7d = Column(Integer, nullable=True)
 
@@ -42,7 +42,7 @@ class ActionSnapshot(Base):
     # Delta results
     delta_cvr = Column(Float, nullable=True)
     delta_atc_rate = Column(Float, nullable=True)
-    delta_revenue_7d = Column(Float, nullable=True)
+    delta_revenue_7d = Column(Numeric(18, 2), nullable=True)
     delta_visitors_7d = Column(Integer, nullable=True)
     delta_orders_7d = Column(Integer, nullable=True)
     delta_computed_at = Column(DateTime, nullable=True)
